Typist/requirements.txt
flask==3.0.3
google-generativeai==0.8.1
langchain==0.2.14
langchain-core==0.2.39
langchain-community==0.2.12
pandas==2.0.2